Add files via upload
In this project, I utilized Vue.js for the frontend and Golang for the backend. During my training, I honed my skills in both technologies, resulting in the development of a medical shop application. Notably, I crafted a visually appealing user interface using Vuetify 2, while all the essential logic was implemented in Golang.In this project , Frontend is created using Vue ,Backend is created using Golang